Add another class forthe error message(abstartct class)
